requirements.txt
# Django and web framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-htmx==1.17.2
django-bootstrap5==23.2
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-filter==23.5
django-import-export==3.3.5
django-allauth==0.54.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Payment integration
stripe==7.8.0
python-mpesa==0.2.8
requests==2.31.0
cryptography==41.0.7

# Email and notifications
django-ses==3.5.0
resend==2.0.0

# File storage and media
django-storages==1.14.2
Pillow==10.1.0

# Development and debugging
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Authentication and security
social-auth-app-django==5.4.0
python-social-auth==0.3.6
social-auth-app-django==5.4.0

# Maps and location
django-leaflet==0.29.0

# Calendar and scheduling
django-scheduler==0.9.0

# Analytics and reporting
django-chartjs==2.2.1
plotly-dash==2.14.2

# Background tasks
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Monitoring and logging
django-health-check==3.17.0

# API documentation
drf-spectacular==0.26.5

# Environment management
python-decouple==3.8
python-dotenv==1.0.0

# Data analysis
pandas==2.1.3
numpy==1.25.2

# Excel/CSV processing
openpyxl==3.1.2
xlrd==2.0.1

# Frontend assets
django-compressor==4.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0
django-heroku==0.3.1

# JavaScript/CSS compilation
nodeenv==1.8.0

# API integrations
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Mobile and SMS
twilio==8.10.0